+++ src/Yaml/ParserInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+/* this file is part of pipelines */
+
+namespace Ktomk\Pipelines\Yaml;
+
+interface ParserInterface
+{
+    /**
+     * @param string $path
+     * @return null|array
+     */
+    public function parseFile($path);
+
+    /**
+     * @param string $buffer
+     * @return null|array
+     */
+    public function parseBuffer($buffer);
+}