+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Support for PHP YAML extension, is preferred over Spyc lib if
+  available; highly recommended
+
 ## [0.0.15] - 2018-04-23
 ### Added
 - Add `--no-dot-env-files` and `--no-dot-env-dot-dist` flags to